requirements.txt
Flask-Cors==4.0.1
Flask-Pydantic==0.12.0
Flask-SQLAlchemy==3.1.1
flask-swagger-ui==4.11.1
mysql-connector-python==8.4.0
pytest==8.2.1
python-dotenv==1.0.1
